+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export { default as Digits } from './Digits';
+export { default as Operation } from './Operation';
